refactor: reorganize project structure, split signIn and SignUp forms into components
