+++ src/app/App.js
@@ -1,0 +1,6 @@
+import React from "react";
+import AppNavigator from "./AppNavigator";
+
+export default function App() {
+  return <AppNavigator />;
+}
