+++ src/features/userAuth/signUp/index.js
@@ -1,0 +1,4 @@
+export { default as SignUpForm } from "./ui/SignUpForm";
+// Если бы были другие экспорты из этой фичи (например, хуки, сервисы), они были бы здесь
+// export * from './model/...'
+// export * from './api/...'
