+++ src/pages/auth/signInScreen.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { SignInForm } from "../../features/userAuth/signIn/ui/SignInForm";
+
+const SignInScreen = () => {
+  return <SignInForm />;
+};
+
+export default SignInScreen;
