+++ src/shared/components/Button.js
@@ -1,0 +1,33 @@
+import React from "react";
+import { TouchableOpacity, Text, StyleSheet } from "react-native";
+import PropTypes from "prop-types";
+
+const Button = ({ onPress, text }) => {
+  return (
+    <TouchableOpacity style={styles.button} onPress={onPress}>
+      <Text style={styles.buttonText}>{text}</Text>
+    </TouchableOpacity>
+  );
+};
+
+Button.propTypes = {
+  onPress: PropTypes.func.isRequired,
+  text: PropTypes.string.isRequired,
+};
+
+const styles = StyleSheet.create({
+  button: {
+    width: "100%",
+    height: 50,
+    backgroundColor: "rgb(29, 29, 29)",
+    justifyContent: "center",
+    alignItems: "center",
+    borderRadius: 10,
+  },
+  buttonText: {
+    color: "#fff",
+    fontSize: 18,
+  },
+});
+
+export default Button;
